Add design doc for "rusty roguelike."
For chapter 4, the task was largely just to write a design doc for the
game we'll build in the next few chapters.

I took most of the text presented in the book as-is and formatted it in
a markdown doc.

I converted the MVP and Stretch Goals sections into checklists in lieu
of ticketing the work out as github issues or something. Not sure if
it'll be useful, but that's the rationale for now.